requirements.txt
requests>=2.24.0
aiohttp>=3.6.2
avro-python3>=1.10.0
jsonschema>=3.2.0
pytz>=2020.1
pbr>=5.5.0
six>=1.15.0
responses>=0.10.15
setuptools>=47.3.1
pydantic>=1.5.1
avro-json-serializer==1.0.3
avro-python3>=1.10.0
aiohttp-sse-client>=0.2.0
janus>=0.5.0
streammachine-schemas-common==0.0.1
